requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
google-generativeai==0.3.2
urllib3==2.0.7
python-dotenv==1.0.0
tqdm==4.66.1
